% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{tgs_finite}
\alias{tgs_finite}
\title{Checks whether all the elements of the vector are finite}
\usage{
tgs_finite(x)
}
\arguments{
\item{x}{numeric or integer vector or matrix}
}
\value{
'TRUE' if all the elements of 'x' are finite, otherwise 'FALSE'.
}
\description{
Checks whether all the elements of the vector are finite.
}
\details{
'tgs_finite' returns 'TRUE' if all the elements of 'x' are finite numbers.
(See: 'is.finite'.)
}
\examples{

tgs_finite(1:10)
tgs_finite(c(1:10, NaN))
tgs_finite(c(1:10, Inf))

}
\keyword{~finite}
