% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tgstat-package.R
\docType{package}
\name{tgstat-package}
\alias{tgstat-package}
\alias{tgstat}
\title{Tanay's group statistical utilities}
\description{
'tgstat' package is intended to help users to efficiently analyze data in
time-patient space.
}
\details{
For a complete list of help resources, use \code{library(help = "tgstat")}.

More information about the options can be found in 'User manual' of the
package.
}
\keyword{package}
