\name{tgstat-package}
\alias{tgstat-package}
\alias{tgstat}
\docType{package}
\title{
Tanay's group statistical utilities
}
\description{
'tgstat' package is intended to help users to efficiently analyze data in time-patient space.

For a complete list of help resources, use \code{library(help = "tgstat")}.

More information about the options can be found in 'User manual' of the package.

}
\keyword{ package }
